% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextCloud4R.R
\docType{package}
\name{nextCloud4R}
\alias{nextCloud4R}
\alias{nextCloud4R-package}
\title{Interface to 'Nextcloud' REST API}
\description{
Provides an Interface to 'Nextcloud' (<https://owncloud.com/>) REST API.
}
\details{
\tabular{ll}{
   Package: \tab nextCloud4R\cr
   Type: \tab Package\cr
   Version: \tab 0.0.9000\cr
   Date: \tab 2019-11-06\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
