% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextCloudManager.R
\docType{class}
\name{nextCloudManager}
\alias{nextCloudManager}
\title{nextCloudManager}
\format{\code{\link{R6Class}} object.}
\usage{
nextCloudManager
}
\value{
Object of \code{\link{R6Class}} for modelling an nextCloudManager
}
\description{
nextCloudManager
}
\section{General Methods}{

\describe{
 \item{\code{new(url, user, pwd, logger)}}{
   This method is used to instantiate an nextCloudManager. The user/pwd are
   mandatory in order to connect to 'nextCloud'. The logger can be either
   NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl 
   http calls logs)
 }
 \item{\code{connect()}}{
   A method to connect to 'Nextcloud' 
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{manager}
\keyword{nextCloud}
