% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextCloudRequest.R
\docType{class}
\name{nextCloudRequest}
\alias{nextCloudRequest}
\title{nextCloudRequest}
\format{\code{\link{R6Class}} object.}
\usage{
nextCloudRequest
}
\value{
Object of \code{\link{R6Class}} for modelling a generic 'nextCloud' web-service request
}
\description{
nextCloudRequest
}
\note{
Abstract class used internally by \pkg{nextCloud4R}
}
\section{Methods}{

\describe{
 \item{\code{new(type, url, request, user, pwd, logger)}}{
   This method is used to instantiate a object for doing an 'nextCloud' web-service request
 }
 \item{\code{getRequest()}}{
   Get the request payload
 }
 \item{\code{getRequestHeaders()}}{
   Get the request headers
 }
 \item{\code{getStatus()}}{
   Get the request status code
 }
 \item{\code{getResponse()}}{
   Get the request response
 }
 \item{\code{getException()}}{
   Get the exception (in case of request failure)
 }
 \item{\code{getResult()}}{
   Get the result \code{TRUE} if the request is successful, \code{FALSE} otherwise
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{nextcloud}
\keyword{request}
